Gkuer

See merge request s06-webmobile2-sub2/S06P12B206!14

+++ back/together-homt/src/main/java/com/ssafy/togetherhomt/feed/FeedDto.java
@@ -27,7 +27,7 @@ public class FeedDto {
 
     private String media_url;
 
-    private Integer like_cnt;
+    private Long like_cnt;
 
     private String userName;
 }
